Correct movie poster displays for movies in database.
Optimized poster search by using title and year:
-Reduced the number of results returned by the API
-Returned results that are more specific

The original method returned the poster url of the first result.
This worked most of the time, but for some movies this returned the incorrect poster.
This change reduced loading time and increased the probability the movie poster displayed would be correct.